More Rubocop Fixes
This actually makes breaking changes to the public methods on list.
As we're not v1.0.0 yet I think this is legit, but may be a good time
to look at marking this stable.

It's an annoying adjustment, but shorter methods make sense and I think
I can afford to bow to convention whilst (I... err... assume..?) I'm the
only one using this gem right now.